Fix to run in java 8
An error occurred in the sorting process when executing the unmodified
script with Astah executed by java 8. I modified the surroundings, but I
changed the specification slightly at that time. Hierarchical indexes
and branch messages will not be in the correct order if you get the
message order from the index. We changed to obtain position information
etc from presentation.